+++ frontend/src/components/setting/Setting.vue
@@ -6,7 +6,7 @@
               <ul class="list-group">
                 <li class="list-group-item" @click="enableMenu('email')" style="cursor: pointer">
                   <b-icon icon="envelope" font-scale="1" aria-hidden="true" class="mx-2"></b-icon>Email Settings</li>
-                <!-- <li class="list-group-item" @click="enableMenu('call')" style="cursor: pointer">
+                <!--<li class="list-group-item" @click="enableMenu('call')" style="cursor: pointer">
                   <b-icon icon="telephone-x" font-scale="1" aria-hidden="true" class="mx-2"></b-icon>Call Settings</li> -->
                 <li class="list-group-item" v-b-modal.modal-1 style="cursor: pointer">
                   <b-icon icon="person-badge" font-scale="1" aria-hidden="true" class="mx-2"></b-icon>Profile Settings
@@ -14,9 +14,6 @@
                 <li class="list-group-item" @click="enableMenu('account')" style="cursor: pointer">
                   <b-icon icon="person" font-scale="1" aria-hidden="true" class="mx-2"></b-icon>Account Settings
                 </li>
-                <!-- <li class="list-group-item" @click="enableMenu('password')" style="cursor: pointer">
-                  <b-icon icon="key" font-scale="1" aria-hidden="true" class="mx-2"></b-icon>Change Password
-                </li> -->
               </ul>
             </div>
             <div v-if="activeMenu == 'email'">